+++ releasenotes.md
@@ -1,4 +1,5 @@
 ## Next version
+* Fixed bug where `AssembliesConfiguration` was never read from `IBootstrapConfiguration`. [#61](https://github.com/ProCoSys/Bifrost/issues/61)
 * More friendly error messages from specifying assemblies. [#59](https://github.com/ProCoSys/Bifrost/issues/59)
 
 ## Version 3.0.2